+++ pipelines/stg/dubinushka_manual/lecturers.sql
