+++ pipelines/stg/mymsuapi/timetable/__init__.py
@@ -121,60 +121,3 @@ def get_timetable_for_semester_to_db():
         index=False,
     )
     return Dataset("STG_MYMSUAPI.raw_timetable_api")
-
-
-@task(
-    task_id="flatten_timetable",
-    inlets=Dataset("STG_MYMSUAPI.raw_timetable_api"),
-    outlets=Dataset("ODS_MYMSUAPI.ods_timetable_api_flattened"),
-)
-def flatten_timetable():
-    sql_engine = sa.create_engine(DB_DSN)
-    sql_engine.execute(
-        """
-        DELETE FROM "ODS_MYMSUAPI".ods_timetable_api_flattened;
-        INSERT INTO "ODS_MYMSUAPI".ods_timetable_api_flattened (
-            group_name,
-            discipline_name,
-            discipline_id,
-            classroom_name,
-            classroom_id,
-            lesson_type_text,
-            lesson_from_dttm_ts,
-            lesson_to_dttm_ts,
-            teacher_full_name,
-            study_group_id,
-            study_group_name
-        )
-        SELECT
-            r.group_name,
-            r.dicscipline_name,
-            r.discipline_id,
-            r.classroom_name,
-            r.classroom_id,
-            r.lesson_type_text,
-            r.lesson_from_dttm_ts,
-            r.lesson_to_dttm_ts,
-            REPLACE(cast(t->'sur_name'as text) || ' ' || cast(t->'first_name' as text) || ' ' || cast(t->'second_name'::text as text), '"', '') AS teacher_full_name,
-            cast(cast(s->'id' as text) as integer) AS study_group_id,
-            s->'name' AS study_group_name
-        FROM "STG_MYMSUAPI".raw_timetable_api r
-        CROSS JOIN LATERAL json_array_elements(r.teacher_users::json) AS t
-        CROSS JOIN LATERAL json_array_elements(r.study_groups::json) as s
-        """
-    )
-    return Dataset("ODS_MYMSUAPI.ods_timetable_api_flattened")
-
-
-with DAG(
-    dag_id="download_mymsuapi_timetable",
-    schedule="50 2 */1 * *",
-    start_date=datetime(2024, 8, 27),
-    tags=["ods", "src", "mymsuapi"],
-    default_args={
-        "owner": "zimovchik",
-        "retries": 3,
-        "retry_delay": timedelta(minutes=5),
-    },
-) as dag:
-    (get_timetable_for_semester_to_db() >> flatten_timetable())
